+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ## 2.7.0
 
 - extend lobby-calling (lobbychanged -> lobby_ambient_in)
-- improve web-caller delay (using indexeddb to cache audio-files)
+- improve web-caller (using indexeddb for caching to continously prevent delays)
 - change file-hoster for more stable downloads (dropbox)
-- update voice-packs (new lobby-sounds, ~3000 new playernames)
+- update voice-packs (new lobby-sounds, 2000+ new playernames)
 - add path argument for blacklist-file
+- fix token-requests
 
 
 ## 2.6.3